+++ nodebackend/src/routes/api.ts
@@ -10,6 +10,7 @@ import switchTaskEnablerRouter from './switchTaskEnablerRoute.js';
 import getTaskEnablerRouter from './getTaskEnablerRoute.js';
 import deleteTaskRouter from './deleteTaskRoute.js';
 import countdownRouter from './countdownRoute.js';
+import et0Router from './et0Route.js';
 
 const router = express.Router();
 
@@ -22,5 +23,6 @@ router.use('/switchTaskEnabler', apiLimiter, switchTaskEnablerRouter);
 router.use('/getTaskEnabler', apiLimiter, getTaskEnablerRouter);
 router.use('/deleteTask', apiLimiter, deleteTaskRouter);
 router.use('/countdown', apiLimiter, countdownRouter);
+router.use('/et0', apiLimiter, et0Router);
 
 export default router;