+++ viteclientts/src/pages/404Page.tsx
@@ -3,7 +3,7 @@ import Layout from '../Layout';
 
 const NotFoundPage: React.FC = () => {
   return (
-    <Layout title="404">
+    <Layout>
       <Grid size={12}>
         <Box sx={{ display: 'flex', flexDirection: 'column', alignItems: 'center' }}>
           <Typography variant="subtitle1">Page not found</Typography>